+++ src/main/java/net/silentchaos512/gear/client/ColorHandlers.java
@@ -1,15 +1,16 @@
 package net.silentchaos512.gear.client;
 
 import net.minecraft.client.renderer.color.ItemColors;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.IItemProvider;
 import net.minecraftforge.client.event.ColorHandlerEvent;
+import net.minecraftforge.registries.ForgeRegistries;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.item.ICoreItem;
 import net.silentchaos512.gear.api.parts.PartType;
 import net.silentchaos512.gear.init.ModItems;
 import net.silentchaos512.gear.item.CustomTippedUpgrade;
-import net.silentchaos512.gear.parts.PartData;
+import net.silentchaos512.gear.item.PartItem;
 import net.silentchaos512.gear.util.GearData;
 import net.silentchaos512.utils.Color;
 
@@ -17,6 +18,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public final class ColorHandlers {
+    @Deprecated
     public static Map<String, Integer[]> gearColorCache = new HashMap<>();
 
     private ColorHandlers() {}
@@ -30,82 +32,81 @@ public final class ColorHandlers {
 
         itemColors.register(CustomTippedUpgrade::getItemColor, ModItems.customTippedUpgrade);
 
-        // Tools/Armor - mostly used for broken color, but colors could be changed at any time
-        itemColors.register(ColorHandlers::getToolLiteColor,
-                ModItems.toolClasses.values()
-                        .stream()
-                        .map(ICoreItem::asItem)
-                        .toArray(Item[]::new)
-        );
-        itemColors.register(ColorHandlers::getArmorLiteColor,
-                ModItems.armorClasses.values()
-                        .stream()
-                        .map(ICoreItem::asItem)
-                        .toArray(Item[]::new)
-        );
-        itemColors.register(ColorHandlers::getShieldColor, ModItems.shield);
+        // Tools, armor, shields, etc.
+        ForgeRegistries.ITEMS.getValues().stream()
+                .filter(item -> item instanceof ICoreItem)
+                .map(item -> (ICoreItem) item)
+                .forEach(item -> itemColors.register(item.getItemColors(), item));
+
+        // Compound part items
+        itemColors.register(PartItem::getColor, ForgeRegistries.ITEMS.getValues().stream().filter(item -> item instanceof PartItem).toArray(IItemProvider[]::new));
     }
 
-    // TODO: When models get fixed, switch back to this
-    @SuppressWarnings("unused")
-    public static int getGearLayerColor(ItemStack stack, int tintIndex) {
-        if (!(stack.getItem() instanceof ICoreItem) || tintIndex < 0) return Color.VALUE_WHITE;
-
-        String modelKey = GearData.getCachedModelKey(stack, 0);
-        Integer[] colors = gearColorCache.get(modelKey);
-        return colors != null && tintIndex < colors.length ? colors[tintIndex] : Color.VALUE_WHITE;
+    /**
+     * Standard colors for most tools, melee weapons, and ranged weapons
+     *
+     * @param stack     The item
+     * @param tintIndex The tint index
+     * @return The color of the layer
+     */
+    public static int getToolColor(ItemStack stack, int tintIndex) {
+        switch (tintIndex) {
+            case 0:
+                return GearData.getColor(stack, PartType.ROD);
+            case 1:
+                return GearData.getColor(stack, PartType.GRIP);
+            case 2:
+                return GearData.getColor(stack, PartType.MAIN);
+            case 3:
+                return GearData.getColor(stack, PartType.TIP);
+            default:
+                return Color.VALUE_WHITE;
+        }
     }
 
-    // Temporary rendering solution until Forge fixes their stuff
+    @Deprecated
     public static int getToolLiteColor(ItemStack stack, int tintIndex) {
-        if (tintIndex == 0) {
-            // Rod
-            PartData part = GearData.getPartOfType(stack, PartType.ROD);
-            if (part == null) return Color.VALUE_WHITE;
-            return part.getFallbackColor(stack, 0);
-        }
-        if (tintIndex == 1) {
-            // Grip
-            PartData part = GearData.getPartOfType(stack, PartType.GRIP);
-            if (part == null) return Color.VALUE_WHITE;
-            return part.getFallbackColor(stack, 0);
-        }
-        if (tintIndex == 2) {
-            // Main
-            return GearData.getHeadColor(stack, true);
-        }
-        if (tintIndex == 3) {
-            // Tip
-            PartData part = GearData.getPartOfType(stack, PartType.TIP);
-            if (part == null) return Color.VALUE_WHITE;
-            return part.getFallbackColor(stack, 0);
-        }
-        return Color.VALUE_WHITE;
+        return getToolColor(stack, tintIndex);
     }
 
+    /**
+     * Standard colors for armor items
+     *
+     * @param stack     The item
+     * @param tintIndex The tint index
+     * @return The color of the layer
+     */
+    public static int getArmorColor(ItemStack stack, int tintIndex) {
+        switch (tintIndex) {
+            case 0:
+                return GearData.getColor(stack, PartType.MAIN);
+            case 1:
+                return GearData.getColor(stack, PartType.TIP);
+            default:
+                return Color.VALUE_WHITE;
+        }
+    }
+
+    @Deprecated
     public static int getArmorLiteColor(ItemStack stack, int tintIndex) {
-        if (tintIndex == 0) {
-            // Main
-            return GearData.getHeadColor(stack, true);
-        }
-        if (tintIndex == 1) {
-            // Tip
-            PartData part = GearData.getPartOfType(stack, PartType.TIP);
-            if (part == null) return Color.VALUE_WHITE;
-            return part.getFallbackColor(stack, 0);
-        }
-        return Color.VALUE_WHITE;
+        return getArmorColor(stack, tintIndex);
     }
 
-    private static int getShieldColor(ItemStack stack, int tintIndex) {
-        if (tintIndex == 0) {
-            PartData part = GearData.getPartOfType(stack, PartType.ROD);
-            if (part == null) return Color.VALUE_WHITE;
-            return part.getFallbackColor(stack, 0);
+    /**
+     * Standard colors for shields
+     *
+     * @param stack     The item
+     * @param tintIndex The tint index
+     * @return The color of the layer
+     */
+    public static int getShieldColor(ItemStack stack, int tintIndex) {
+        switch (tintIndex) {
+            case 0:
+                return GearData.getColor(stack, PartType.ROD);
+            case 1:
+                return GearData.getColor(stack, PartType.MAIN);
+            default:
+                return Color.VALUE_WHITE;
         }
-        if (tintIndex == 1) {
-            return GearData.getHeadColor(stack, true);
-        }
-        return 0xFFFFFF;
     }
 }