+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Some changes to how colors are handled. All layer colors are now cached in NBT. Locked and unlocked gear should update.
+- (API) `ICoreItem` now has a method to return an `IItemColor`, which will be automatically registered. This defaults to the appropriate method for most items.
+
 ## [1.7.2] - 2020-06-05
 ### Added
 - Moar new traits!